+++ packages/blackbox_core/lib/src/test.dart
@@ -7,7 +7,7 @@ Cancel cancelGuardedForTest(void Function() fn) => _cancelGuarded(fn);
 // graph.dart
 @visibleForTesting
 T resolveDependencyForTest<T>(Graph g, OutputSource<T> box) {
-  return DependencyResolver._(g).whenReady(box);
+  return DependencyResolver._(g).onlyWhenReady(box);
 }
 
 @visibleForTesting